+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Configurazioni deploy Semaphore
 + Fix semaphore
 + Implementazione function randomNumber e aggiunta coverAll
++ Implementazione DashboardPage + test
 
 ## 0.1.0
 + Aggiungo roba